refacrtor queries
